+++ classifai-api/pom.xml
@@ -30,7 +30,6 @@
         <dependency>
             <groupId>ai.classifai</groupId>
             <artifactId>classifai-core</artifactId>
-            <version>2.0.0</version>
         </dependency>
         <dependency>
             <groupId>io.vertx</groupId>
