+++ classifai-api/src/main/java/ai/classifai/config/PortSelector.java
@@ -15,7 +15,7 @@
  */
 package ai.classifai.config;
 
-import ai.classifai.server.ParamConfig;
+import ai.classifai.util.ParamConfig;
 import lombok.NonNull;
 import lombok.extern.slf4j.Slf4j;
