+++ classifai-core/src/main/java/ai/classifai/database/annotation/boundingboxdb/BoundingBoxVerticle.java
@@ -17,10 +17,9 @@ package ai.classifai.database.annotation.boundingboxdb;
 
 import ai.classifai.database.DatabaseConfig;
 import ai.classifai.database.annotation.AnnotationVerticle;
-import ai.classifai.server.ParamConfig;
+import ai.classifai.util.ParamConfig;
 import ai.classifai.util.message.ErrorCodes;
-import ai.classifai.annotation.AnnotationType;
-//import ai.classifai.util.type.AnnotationType;
+import ai.classifai.util.type.AnnotationType;
 import io.vertx.core.Promise;
 import io.vertx.core.eventbus.Message;
 import io.vertx.core.json.JsonObject;
