+++ classifai-core/src/main/java/ai/classifai/database/portfoliodb/PortfolioDbQuery.java
@@ -44,8 +44,6 @@ public class PortfolioDbQuery
 
     private final static String GET_PROJECT_UUID_LIST = "select uuid_list, uuid_generator_seed from Portfolio where project_id = ?";
 
-    //private final static String REMOVE_OBSOLETE_UUID_LIST = "Removal of obsolete uuid";
-
     private final static String GET_ALL_PROJECTS_FOR_ANNOTATION_TYPE = "select project_name from Portfolio where annotation_type = ?";
 
     public static String getQueue(){ return QUEUE; }
@@ -68,7 +66,7 @@ public class PortfolioDbQuery
 
     public static String getProjectIDList() { return GET_PROJECT_ID_LIST; }
 
-    public static String getProjectUUIDListt() { return GET_PROJECT_UUID_LIST; }
+    public static String getProjectUUIDList() { return GET_PROJECT_UUID_LIST; }
 
     public static String getAllProjectsForAnnotationType() { return GET_ALL_PROJECTS_FOR_ANNOTATION_TYPE; }
