+++ keyShortcut.md
@@ -1,0 +1,33 @@
+# Classifai Keyboard Shortcut 
+## Introduction
+
+The interface of Classifai is user-friendly to use and very easy to master it. For make user's task more easy to handle, 
+these are few common and useful shortcut keys have been created. 
+
+## 1.0 Bounding Box
+
+![Bounding-box-keyboard](/metadata/bounding-box-keyboard.png)
+
+| No | Shortcut Key | Functionality | 
+|----| -------------|-------------|
+| 1  | Ctrl + alt   | Toggle green light on/off <ul><li> **Green Light on**: move a selected boundingbox by using arrow keys (right, left, up, down)</li> <li>**Green light off:** Go to next/previous images by using left/right key</li></ui> |
+| 2  | Ctrl + shift | Toggle between drawing bounding box & moving image| 
+| 3  | Ctrl + z     | Undo the last action |
+| 4  | Ctrl + c     | Copy a selected bounding box|
+| 5  | Ctrl + v     | Paste a selected bounding box |
+| 6  | Delete/Backspace | Delete a selected bounding box |
+| 7  | Double click on a selected bounding box in the UI | Choose label for the selected bounding box |
+
+## 2.0 Segmentation
+![Segmentation-keyboard](/metadata/segmentation-keyboard.png)
+
+| No | Shortcut Key | Functionality |
+|----| -------------|-------------|
+| 1  | Ctrl + alt   | Toggle green light on/off <ul><li>**Green light on:** move a selected polygon by using arrow key</li> <li>**Green light off:** left/right key to change to next/previous images</li></ul> |
+| 2  | Ctrl + shift | Toggle between drawing Polygon & moving image | 
+| 3  | Ctrl + z     | Undo the last action |
+| 4  | Ctrl + c     | Copy a selected polygon |
+| 5  | Ctrl + v     | Paste a selected polygon |
+| 6  | Delete/Backspace | <ul><li>Delete a selected polygon</li> <li>Delete the last drawing dot (during drawing new polygon)</li></ul> |
+| 7  | Double click on a selected polygon in the UI | Choose label for the selected polygon |
+| 8  | Esc          | Delete the polygon currently in the midst of drawing |
