+++ startClassifai.bat
@@ -1,7 +1,7 @@
 @REM Copyright (c) 2020 CertifAI Sdn. Bhd.
 
 @echo on
-SET JAR="%userprofile%\.m2\repository\ai\classifai\classifai-uberjar\1.0.0\classifai-uberjar-1.0.0-dev.jar"
+SET JAR="%userprofile%\.m2\repository\ai\classifai\classifai-uberjar\2.0.0\classifai-uberjar-2.0.0-dev.jar"
 
 del %JAR%
